01_concatmath/pyproject.toml
[project]
name = "01-concatmath"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core>=0.12.44",
    "llama-index-llms-deepseek>=0.1.2",
    "llama-index-utils-workflow>=0.3.5",
]
